fix: rename xsiam_interface.py to xsiam_client.py
